Separación de clases sobre versión geat/scaffold
