+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 group = 'us.mitfs.samples'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '17'
+sourceCompatibility = '16'
 
 repositories {
 	mavenCentral()
